ingestor/pyproject.toml
[project]
name = "ingestor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
]
